benchmarking: Clone Gitaly using target revision
Currently we clone Gitaly at HEAD and then peel the requested revision
and check that out. This works fine if you're using HEAD as your rev,
but a branch name will fail as we haven't created a local branch with
that name yet.

To resolve this, perform the initial clone using the requested revision
before peeling.