Default enable invalid metadata deletion in Praefect
Praefect has a metadata verifier that checks whether the replicas
recorded in the database actually exist. For now, it has only logged
and gathered metrics on invalid metadata records by default. The
verifier also supports deleting invalid metadata which clean up the
stale records and allows Praefect to realize the replicas are missing
and trigger replication. This functionality has been disabled in the
past as renames racing with the verificaton could cause data loss.
Since 15.3, Praefect has been using Praefect generate replica paths
and not performing any more renames on disk. Given this, there are
no more races that we know of, so we can default enable the invalid
metadata deletion.

Changelog: changed