gittest: Respect object format when creating repos via the service
When creating repositories via `gittest.CreateRepository()` then the
caller has the ability to pass in a repository format that shall be used
for new empty repositories. And while we respect the format when we
initialize the repository directly, we don't pay attention to it when we
create the repository via the `CreateRepository()` RPC call.

Fix this by correctly passing the requested object format to the RPC's
request.