praefect: Fix test for feature flag transactions
Currently the praefect `TestDisabledTransactionsWithFeatureFlag` test
only validates one of the RPCs then breaks out of the loop. If the
`RestoreCustomHooks` RPC got randomly selected as the RPC to be tested,
the test would panic because the context would not contain the required
feature flag. This change removes the break allowing all RPCs to be
validated and also adds the feature flag to the test context to prevent
the panic.