git: Include reason oid is invalid in error
Previously `ValidateHex` returned a generic error message when an
invalid string representation of a SHA1 or SHA256 oid was found. This
gave no indication as to why the oid was rejected.

Now that we are iterating over characters manually rather than using a
regexp, we are in a position to provide users with more detail on what
is invalid about the object ID. Introduce new error types to bubble this
information up to users.