ruby: Match gem versions of gitlab-org/gitlab
To reduce the number of gem installed in a complete GitLab installation
it's nice to have all components use the same gem versions. This change
changes the currently mismatches versions with those of
gitlab-org/gitlab.

As an added benefit, this effort ensures all gems are compatible with
Ruby 3, as gitlab-org/gitlab updated all gems in [1].

[1]: https://gitlab.com/groups/gitlab-org/-/epics/6210