operations: Fix error paths in CherryPick on custom hook error
While checking whether we can remove the `PreReceiveError` field of the
`UserCherryPick` response I stumbled upon one branch where we seemingly
still setting that field. As it turns out though we don't as the error
condition is accidentally handled twice, and as a result we would always
return the structured error as expected.

Fix the confused code by removing the redundant-and-wrong error
handling.