tracing: Fix span name and contents in command generator
The span attached to shell-out command used the full path as the
operation name. It makes the traces of this type noisy. In addition, it
attached the command result as logs. Depending on the platform, those
logs are usually detached from the span. This commit improves the span
name and treat command results as tag values.