tracing: Fix orphaned spans from catfile cache
Previously, we create a span before issuing catfile.getOrCreateProcess.
This span finishes when the process exists. This process cache intends
to share the process between requests. Its lifecycle spans between
different requests, and may last for several seconds/minutes. As a
result, the original span never ends. Eventually, it is sent to tracing
server and becomes an orphaned span. This situation adds a lot of
noises.

This commit changes the semantic of catfile.* span. It is now the span
of requesting an object reader. The span finishes as soon as the caller
gets the reader as it wishes.