gitaly: Add `deleteKey` operation to `TransactionManager`
When we apply an operation from the write-ahead log, we also want to
remove that entry from the log. Before we add this logic, it is
necessary that `TransactionManager` supports deletion.

This commit modifies the `databaseTransaction` interface to add a
`Delete` function and also adds a `deleteKey` function to the
`TransactionManager`.