ci: Allow cancellation of redundant pipelines
Pipelines that are made redundant by a subsequent push are not getting
cancelled even though we've enabled the corresponding setting in the
Gitaly project. Seemingly, this is a feature that was introduced in
GitLab 12.3 where jobs also need to declare that they are interruptible
or otherwise GitLab won't honor that setting.

Mark all of our jobs as interruptible to stop wasting compute.