repository: Replicate custom hooks
Currently the `ReplicateRepository` RPC does not include the replication
of custom hooks for a repository. This change updates the RPC to include
replication of repository hooks stored in the `custom_hooks` directory
when the `gitaly_replicate_repository_hooks` feature flag is enabled.
The hooks replication is handled by `setCustomHooksTransaction()` from
the `SetCustomHooks` RPC which enables atomicity and transactionality.