Gitaly: Refactor ConfigureRuby()
The 'ConfigureRuby()' method now returns 'ValidationErrors'.
The check were re-arrange a bit to group them together.
Also, the type of the 'MaxRSS' and 'NumWorkers' changed
to accept only unsigned values as negative values are not
allowed for these settings. And the checks for the duration
were added for 'GracefulRestartTimeout' and 'RestartDelay'.