datastore: Increase timeout for storage cleanup test
We've recently seen one of our tests for storage cleanup flake because
it hit the timeout of two seconds. This isn't all that surprising on
machines that are as slow as our CI runners.

Bump the timeout to a full minute. While it would be very much
preferable to refactor the code so that it doesn't need any timeouts in
the first place it doesn't feel worth it to refactor the code now.