updateref: Unify error handling for I/O errors
Error handling when reading from or writing to the git-update-ref(1)
process needs to be asynchronous because the process will not respond to
all the commands we're sending. Consequentially, it can happen that an
error caused by a write will be surfaced by a later read. We thus have
logic in place both for write and read paths that handles the common
error handling.

There still is some duplicated logic around though that involves killing
the process itself so that we can process its standard error in the
first place. Unify it so that we can extend it more readily without
having to increase the duplicated code even further.