supervisor: Refactor event notifier
The supervisor package has functionality to notify listeners when
the supervised process causes a certain event to happen. This logic
performs the notification asynchronously so that it can exit if there
was no receiver for the notification after at most one second. Given
that this mechanism shouldn't block the spawn loop all callers of that
notifier spawn it in a separate Goroutine.

Refactor the code so that the Goroutine gets spawned as part of the
notifying function itself.