Gitaly: validate-configuration sub-command
In order to improve developers experience in using Gitaly
the new 'validate-configuration' is added. The purpose of
it to validate provided configuration before starting
the service. The output lists all the problems of the
configuration in JSON format into STDOUT. The structure
of object includes 'key' which is a path to the field
where the problem detected and the 'message' with an
explanation of the problem.

Changelog: added

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/4650