dnsresolver: Remove retry assertion in test
Dnsresolver tests attempt to assert the number of retries in the
exponential backoff provider. It looks like an overkilled assertion.
Plus, DNS queries via UDP may fail, leading to unexpected retries
outside of the test’s intention. We should better remove that assertion.
As soon as the final result is the same, we don’t care how many times
the resolver retries.