dnsresolver: Remove random delay in tests
In a prior MR, we refactored the fake connection so that it blocks the
resolver when it received enough data. As a result, we don’t need a 1ms
delay anymore. The fake connection creates as a natural backpressure.
This commit sets the duration in test to 0 to disable the timer.