git/packfile: Recognize packfiles in SHA256 repositories
Packfiles in Git repositories that use the SHA256 object format will
also use SHA256 as part of their name. The regular expression we have in
our `packfile` package to filter packfiles only recognizes ones with
exactly 40 hex characters though, so it fails to identify packfiles when
SHA256 is in use.

Adapt the regular expression to also detect packfiles with the SHA256
hash. Refactor the tests to use `gittest.CreateRepository()` so that we
can automatically test with SHA256 object format.