go: Fix dependencies which have been inadvertently downgraded
With 8a29ef511 (Implement basic transaction processing with write-ahead
logging, 2022-10-14) we have implemented the basic logic for write-ahead
logging of references. As this MR introduced a bunch of logic it has
been taking a comparatively long time to get it reviewed and landed. As
part of that there had seemingly be a merge conflict in the `go.mod`
file that was mis-merged and that caused us to downgrade the following
dependencies:

    - github.com/getsentry/sentry-go 0.17.0 => 0.16.0
    - github.com/rubenv/sql-migrate 1.3.0 => 1.2.0
    - golang.org/x/sys 0.4.0 => 0.3.0
    - golang.org/x/time 0.3.0 => 0.2.0

Revert these downgrades so that we're using the intended dependency
versions again.

Changelog: fixed