ssh: Remove unneeded cleanup function returned from `setupSSHClone`
The `setupSSHClone()` helper sets up a new Git repository that we'll use
as the source of a push. It then returns both the details of that repo
as well as a cleanup function to delete both the created repo's path as
well as the repo we're ultimately pushing into.

Given that we write these repositories into temporary directories that
get cleaned up at the end of the test anyway it is not necessary to
invoke the cleanup function manually. Remove it.