ci: Don't specify minor version for Postgres
We're currently specifying the minor version that is to be used for the
Postgres service that we have running for our Praefect tests. This feels
overly specific as what we really care about is the major version.

Change the image tags to just specify the major version so that we pull
in new bugfix releases automatically.