remote: Have FindRemoteRepository detect repos using SHA256
The FindRemoteRepository RPC can be used to detect whether a repository
exists at the given URI. This is done by asking git-ls-remote(1) to
enumerate the remote's HEAD and then checking that we recognize it as a
valid reference.

This check breaks with the introduction of the SHA256 object format as
we explicitly verify that HEAD points to an object ID that is 40 hex
characters long. So when seeing a 64 hex characters long SHA256 hash
then we will wrongfully return that the repository does not exist.

Now this whole check feels kind of weird, and specifically verifying
specific object hash lengths isn't a great way to figure out whether a
repository exists or not. But for the time being we don't really have a
better way to figure this out, so our hands are kinda tied.

Refactor the RPC to recognize both SHA1 and SHA256 object hashes. While
this doesn't make the underlying issue any better, it at least continues
to work the sahe with repositories that use the SHA256 object format.

Changelog: added