Praefect: Failover.Validate method
The new 'Validate' method validates values of the 'Failover' type and
returns all found errors together. If 'Enabled' field is set to
'false' no validation done as it won't be used and doesn't affect
the service anyhow. It will be used in the later changes.

The cfgerror now has a new 'IsSupportedValue' check.