proto: Add `WriteCustomHooks` RPC
Currently the `RestoreCustomHooks` RPC is used to set the custom git
hooks for a repository from a backup. In the future this RPC will be
used more generally to write hooks to a specified repository. To
accommodate this, this change implements a new RPC `WriteCustomHooks`
which performs the exact same operation but which a more general name.