proto: Add `GetCustomHooks` RPC
Currently the `BackupCustomHooks` RPC is used during backups to fetch
the custom git hooks within a repository and returns them as a tarball.
In the future, this RPC will be used more generally to retrieve custom
hooks from a specified repository. This change implements the
`GetCustomHooks` RPC which performs the same operations but has a more
general name for semantic purposes.