localrepo: Allow unrelated histories in merge tree
In Git, by default git-merge(1) and git-merge_tree(1) throw an error
when an attempted merge is made on two commits that do not have a shared
ancestor. There are times when we want to allow this. For example, if a
project has a default branch `master` that has some history in it, then
the project decides to create a new branch `main` and wants to merge
`master` back into `main`, the two branches will not have shared
history.

Create an option that allows these kinds of merges.