operations: Allow unrelated histories in UserMergeBranch
UserMergeBranch will receive requests to merge commits with unrelated
histories. This commit allows that in the RPC.

7188fee10 (2023-11-15: operations: Implement merge() using git-merge-tree(1))
modified the `TestUserMergeBranch_conflict` test to add a common
ancestor. This was incorrect, as we should not have modified existing
tests to match the new implementation. This commit restores the original
test before 7188fee10 so the two commits being merged do not share a
common ancestor.