gitaly: Validation of Git configuration
We want to make errors of configuration validation more verbose
and flexible to operate with. The idea is to produce a machine
friendly output that is why the ValidationError structure is defined.
The next steps are to refactor existing validation methods to return
errors as instances of the ValidationError combined into
ValidationErrors.

The first refactored validation is for the Git type. It now has a
new Validate method that returns all the validation errors found.