gitaly: Apply hooks on each step
The `TestTransactionManager` only adds hooks when it first creates the
manager. This means if the manager was created in step X, the hooks from
step X+1 won't be added unless the manages is stopped/started. To avoid
this issue, we create a new `applyHooks` function and call this at the
beginning of each step.