gitaly: Add more information to `hookContext`
The `hookContext` structure is provided to the handler of every
synchronization points. When we introduce the deletion of entries in the
write-ahead log, we want to check the state of the database and
pass/fail the test. To do this, we need the {database, test} structs.
Let's add them to the `hookContext`.

We override the default `testing.TB` object with our custom
`testingHook` struct, which overrides calls to `FailNow()` with calls to
`Fail()`, since the former can only be run in the main goroutine. This
allows us to now not worry about using `FailNow()` (this is the function
called by the `require` library).