gitaly: Prune log entry after being applied
In `TransactionManager`, we don't Prune a log entry after it is
applied. This means the key lies in the database forever while providing
no additional value. Let's start pruning entries in the log after being
applied, we do this in `TransactionManager.applyLogEntry` wherein after
the transaction is applied we go ahead and delete the entry.

Since the entries are deleted, some of the tests need modification to
check the state of the database before deletion, modify all such tests.