middleware: Change method lookup warning log to debug
This log line is outputted when the given method cannot be found in
praefects method registry. This registry has definitions and metadata
for every method that praefect proxies for gitaly.

This error would only ever be significant to developers that are adding
more RPCs to gitaly/praefect. Hence this log should actually be debug
output, not warn. This warning only confuses end users who think that
something has gone wrong.