tracing: Implement some tracing ultility methods for Gitaly
Tracing is now powered by opentracing and labkit. They work fine, but
lack of some functionality, such as discarding all spans of path. In
addition, in the future, we may need to switch the tracing library. It
makes sense to implement a thin wrapper providing minimal interfaces.

This commit implements some foundation methods for this purpose.