tracing: Detach git catfile cache's span
Git Catfile cache is an optimization. It re-uses git-cat-file process if
it's already open by other goroutines. As a result, the process's life
span may be stretched outside the request starting the process. At
present, that span is attributed to the first request. That makes the
timing visualization of the whole trace looks weird. This commit
detaches that span from the starting request.