gitaly: Validation of Ruby
The Ruby configuration should be checked before it is used.
The new 'Validate()' method now added, but it is not yet used
to keep the old behaviour. The 'MaxRSS' and 'NumWorkers' field
type changed to uint as the negative values are not valid for them.

Refactor 'Validate()' of the prometheus.Config to use 'IsPositive()'
check.