Implement Git trace2 events parser
Git v2.20 introduces new Trace2 API. This API provides kinda hook to
extract the internal telemetry information of a Git process. It writes
the data to stderr or a file. The events are structured as a flat list
of consecutive events. When GIT_TRACE2_BRIEF variable is set, timing
information is omitted. That makes this format particularly hard to use
later.

In the foreseeable future, we use trace2 in two places: generating
equivalent distributed spans and expose pack object metrics. In the
future, we are looking forward to utilizing Trace2 more. As a result, we
should create an internal representation serving our purposes.
Naturally, the data can be presented as a tree.

This commit creates a parser that transform flat events to a tree data
structure.

For more information: https://git-scm.com/docs/api-trace2