repository: Validate repository exists for hooks
Currently the `GetCustomHooks` and `SetCustomHooks` RPCs do not validate
that the git repository actually exists. This can lead to a confusing
error message when the tar command fails due to the non-existent repo
path. This change updates both RPCs to use `locator.GetRepoPath` which
validates that both the storage and git repository exist.