testhelper: Add AssertGrpcCode helper function
Most of our testhelper functions use the `require` family of assertions,
which internally use use `t.FailNow()`. However, this is not safe for
use goroutines other than the one actually running the test function.
Calling it from a helper goroutine will not determinstically stop the
test and may lead to spurious failures from the race detector.

Add an `AssertGrpcCode` function that uses `assert`, which can safely be
called from a helper goroutine.