gitaly: Validation of Gitlab
The 'Gitlab' configuration should be checked before it is used.
The new 'Validate()' method now added. The same done for the
'HTTPSettings' type. The method is not yet used to keep the
old behaviour.
The type of the 'ReadTimeout' field changed to accept only positive
values as only allowed for it.

The 'cfgerror' package extended with a new 'InRange()' validation
function that is used to verify the value belongs to the expected
range.