gitaly: Add functionality to update default branch
Transactions and log entries in the write-ahead log now have fields for
considering updates to the default branch. Let's add code to consider
these fields and update the default branch.

Most of the heavy lifting is done by `localrepo.SetDefaultBranch`, we
only validate if the refname is either:
1. Is one of the refnames in the current transaction and is not
scheduled to be deleted.
2. Is a existing refname in the current repository.