housekeeping: Unconditionally enable writing multi-pack indices
We have introduced the logic to write multi-pack-indices as part of our
repository housekeeping logic in v15.9. The initial rollout has caused
some issues with pooled repositories, where we thought it was fine to
write bitmaps when multi-pack-indices are enabled. But this turned out
to still not be allowed and was subsequently fixed via c74a4b738
(housekeeping: Don't write bitmaps with MIDX in pooled repos,
2023-01-31). The subsequent rollout to production was successful and
the feature flag was default-enabled in v15.9.

Remove the feature flag guarding this logic. This should allow us to
more efficiently serve larger repositories that have many packfiles.

Changelog: changed