blob: Don't hardcode LFS pointer object IDs
In our LFS-focussed tests we use a set of hardcoded LFS pointer test
data along with hardcoded object IDs. But hardcording the object IDs is
not required anymore as we always write LFS pointer data at test runtime
anyway.

Refactor the test setup to instead derive the object IDs at runtime when
creating the test repository and embed the LFS pointers into the struct
we use to describe the LFS repository.