blob: Enable testing with SHA256 for most tests
All tests of the blob service now use unseeded repositories and should
thus work with both the SHA1 and SHA256 object format. Remove the build
tags to enable the SHA256 tests.

Note that we cannot yet enable tests for the GetBlobs RPC. This RPC
depends on the `catfile.TreeEntries` interface, which does not yet
handle the SHA256 object hash.